refactor(letta-code): Extract result-fetch/usage helpers into letta_code_results
Move the post-run result-extraction helpers off LettaCodeTarget into a new
letta_evals/targets/letta_code_results.py module as free functions:

- extract_usage_stats
- fetch_trajectory
- fetch_token_data
- _run_sort_key (module-private, used by fetch_token_data)

These depend only on the Letta client or a list of stream events, not on
target configuration, so they don't belong on the class. The target keeps its
core job (launch the CLI, assemble a TargetResult) plus the subprocess/stream
helpers. Pure reorganization, no behavior change.

Token-data tests now call fetch_token_data directly instead of constructing a
LettaCodeTarget to reach a helper.
